Split prior weight tests and dedupe solver bucket constants.
Asset parse tests and catalog integration tests belong in separate modules,
and solver tests should import bucket definitions from the canonical module.
